Update_After_Testing
This update is after testing the plugin on a spare Windows Laptop. Debugged issues that were causing the plugin to not load, then debugged issues that were causing the plugin to not give an output. Also fixed the issue regarding files that were not loading upon opening the solution code.